Moved LoggingService to Core namespace and introduced Core header
