+++ samples/EcsPipeline/main.cpp
@@ -11,8 +11,8 @@ NovelRT::Utilities::Event<NovelRT::Timing::Timestamp> DummyUpdateStuff;
 
 int main()
 {
-    NovelRT::LoggingService logger = NovelRT::LoggingService();
-    logger.setLogLevel(NovelRT::LogLevel::Info);
+    NovelRT::Core::LoggingService logger = NovelRT::Core::LoggingService();
+    logger.setLogLevel(NovelRT::Core::LogLevel::Info);
 
     DefaultPluginSelector selector;
     auto windowingProvider = selector.GetDefaultPluginTypeOnCurrentPlatformFor<IWindowingPluginProvider>();
